lgamma gamma_p gammap_inv

+++ test/doc/core/gamma_pinv.cpp
@@ -1,10 +1,10 @@
+#include <iostream>
+#include <iomanip>
 #include <eve/function/gamma_pinv.hpp>
 #include <eve/function/gamma_p.hpp>
 #include <eve/wide.hpp>
-#include <iostream>
-#include <iomanip>
 
-using wide_ft = eve::wide<float, eve::fixed<4>>;
+using wide_ft = eve::wide<double, eve::fixed<4>>;
 
 int main()
 {
